Add drawers in all wood types
This is still fully MineClone 2 compatible; there are different
drawers in MCL2 than MTG and they have different textures.

If you combine different wood types in the crafting field, the
default drawer (oak/wooden) will be crafted. If you use the same
wood, a drawer of this wood will be crafted.