Utilize existing drawers insert function, add pipeworks support, and remove hard depend on default
When I was adding in pipeworks support for the controller, I noticed there functions in the drawers mod already for inserting items in to a drawer. I utilized that to reduce the code and keep things consistent. Pipeworks support was added by looking at how the drawers handled it and looking at the api. I noticed the drawers mod didn't have a hard depend on default so I added in a fallback recipe that way my controller doesn't require the whole mod to have a dependency on default.