Fix empty drawer indexing bug and pipeworks tube connection
There was a flaw in the logic of indexing empty drawers. I thought getting a node metadata string that was never set would return nil, but it actually returned a blank string. This meant drawers that hold more than one item were being treated as if they also had an empty spot that was non-existent. I used an additional node metadata string that the drawers have to distinguish if the drawer has that visual id or not.

Also I forgot to allow pipeworks tubes to connect to the front of the controller.